apps/backend/requirements.txt
# Django & DRF
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-tenants==3.5.0
django-extensions==3.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Celery
celery==5.3.4
redis==5.0.1

# AI & ML
openai==1.6.1
anthropic==0.8.1
langchain==0.1.0
llama-index==0.9.20
pandas==2.1.4
numpy==1.26.2

# Document Processing
python-docx==1.1.0
openpyxl==3.1.2
PyPDF2==3.0.1
pdfplumber==0.10.3
weasyprint==60.2

# Validation & Serialization
pydantic==2.5.2
marshmallow==3.20.1

# Security
djangorestframework-simplejwt==5.3.0
django-otp==1.3.0
cryptography==41.0.8

# Monitoring & Observability
django-prometheus==2.3.1
sentry-sdk==1.39.1
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Feature Flags
unleash-client==5.8.0

# Utilities
python-decouple==3.8
Pillow==10.1.0
requests==2.31.0
boto3==1.34.0
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0
black==23.12.0
isort==5.13.2
flake8==6.1.0

# Stripe
stripe==7.11.0
